user services cleanup
